Renames files as per nestjs convention and adds skeleton   runner-factory service.
