+++ servers/execution/runner/src/app.module.ts
@@ -1,7 +1,8 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
+import RunnerFactoryService from './runner-factory.service.js';
 import AppController from './app.controller.js';
-import ExecaManager from './execaManager.service.js';
+import ExecaManager from './execa-manager.service.js';
 import Queue from './queue.service.js';
 import configuration from './config/configuration.js';
 
@@ -13,6 +14,6 @@ import configuration from './config/configuration.js';
     }),
   ],
   controllers: [AppController],
-  providers: [ExecaManager, Queue],
+  providers: [ExecaManager, Queue, RunnerFactoryService],
 })
 export default class AppModule {}
