+++ servers/execution/runner/src/runner-factory.service.ts
@@ -1,0 +1,4 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export default class RunnerFactoryService {}
