+++ servers/execution/runner/test/integration/execa-runner.spec.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from '@jest/globals';
 import Runner from 'src/interfaces/runner.interface';
-import ExecaRunner from 'src/execaRunner';
+import ExecaRunner from 'src/execa-runner';
 
 describe('check command Runner based on execa library', () => {
   it('should execute a valid command', async () => {
