+++ servers/execution/runner/test/integration/queue.service.spec.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from '@jest/globals';
 import Queue from 'src/queue.service';
 import { Command } from 'src/interfaces/command.interface';
-import ExecaRunner from 'src/execaRunner';
+import ExecaRunner from 'src/execa-runner';
 
 const commands: Command[] = [
   {
